+++ idea/testData/refactoring/rename/classUsagesInTextFiles/before/test/test.kt
@@ -1,0 +1,9 @@
+package test
+
+class A {
+    private class B {
+        class /*rename*/C {
+
+        }
+    }
+}